Get rid of webconsoleUI
