+++ src/devtools/client/inspector/markup/markup-context-menu.js
@@ -320,7 +320,7 @@ class MarkupContextMenu {
   _showDOMProperties() {
     this.toolbox.openSplitConsole().then(() => {
       const { hud } = this.toolbox.getPanel("webconsole");
-      hud.ui.dispatchEvaluateExpression("inspect($0, true)");
+      hud.dispatchEvaluateExpression("inspect($0, true)");
     });
   }
 
@@ -332,7 +332,7 @@ class MarkupContextMenu {
    * autofills it with the temp variable.
    */
   async _useInConsole() {
-    return
+    return;
     await this.toolbox.openSplitConsole();
     const { hud } = this.toolbox.getPanel("webconsole");
