+++ src/ui/components/Toolbox.js
@@ -423,7 +423,7 @@ class Toolbox extends React.Component {
             {...this.getSplitBoxDimensions()}
             splitterSize={1}
             vert={false}
-            onResizeEnd={num => { }}
+            onResizeEnd={num => {}}
             startPanel={
               <div className="toolbox-top-panels">
                 <div
